@@ -1,6 +1,6 @@
 {
   "name": "next-intl",
-  "version": "1.0.1",
+  "version": "2.0.0",
   "private": true,
   "scripts": {
     "dev": "next dev",
@@ -10,18 +10,17 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "gray-matter": "^4.0.2",
-    "next": "^12.1.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "remark": "^14.0.2",
-    "remark-html": "^15.0.1"
+    "gray-matter": "^4.0.3",
+    "markdown-to-jsx": "^7.4.0",
+    "next": "^14.0.4",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0"
   },
   "devDependencies": {
-    "@types/node": "^17.0.14",
-    "@types/react": "^17.0.39",
-    "eslint": "8.8.0",
-    "eslint-config-next": "12.0.10",
-    "typescript": "^4.5.5"
+    "@types/node": "^20.11.0",
+    "@types/react": "^18.2.47",
+    "eslint": "8.56.0",
+    "eslint-config-next": "14.0.4",
+    "typescript": "^5.3.3"
   }
 }
