@@ -1,12 +1,12 @@
-import Navigation from '../Navigation';
+import Navigation from "../Navigation";
 
 type Props = {
   className?: string;
   children?: React.ReactNode;
 };
 
-export default function Layout({ className, children }: Props) {
-  const footerClass = className || 'footer';
+export function Footer({ className, children }: Props) {
+  const footerClass = className || "footer";
 
   return (
     <footer className={footerClass}>
