@@ -1,17 +1,15 @@
-import Link from 'next/link';
+import Link from "next/link";
 
 type Props = {
   link?: string;
 };
 
 export default function Logo({ link }: Props) {
-  const href = link || '/';
+  const href = link || "/";
 
   return (
-    <Link href={href}>
-      <a className="logo">
-        <img src="/svg/logo.svg" alt="logo" />
-      </a>
+    <Link className="logo" href={href}>
+      <img src="/svg/logo.svg" alt="logo" />
     </Link>
   );
 }
